# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediaDrop API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ParseUrlInput {
  /** The media URL to parse */
  url: string;
}

export interface MediaFormat {
  id: string;
  /** Human-readable quality label (e.g. "1080p HD", "MP3 320kbps") */
  label: string;
  /** video, audio, or image */
  type: string;
  extension: string;
  /** @nullable */
  quality?: string | null;
  /**
     * Estimated file size
     * @nullable
     */
  size?: string | null;
  /**
     * Direct download URL if available
     * @nullable
     */
  url?: string | null;
}

export interface MediaInfo {
  url: string;
  /** Detected platform (youtube, instagram, tiktok, etc.) */
  platform: string;
  /** video, audio, image, or carousel */
  mediaType: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  /**
     * Duration in seconds
     * @nullable
     */
  duration?: number | null;
  /** @nullable */
  author?: string | null;
  formats: MediaFormat[];
}

export interface DownloadInput {
  url: string;
  platform: string;
  mediaType: string;
  title: string;
  /** @nullable */
  thumbnail?: string | null;
  /** @nullable */
  author?: string | null;
  /** Quality label selected */
  format: string;
  extension: string;
  /** @nullable */
  fileSize?: string | null;
}

export interface Download {
  id: number;
  url: string;
  platform: string;
  mediaType: string;
  title: string;
  /** @nullable */
  thumbnail?: string | null;
  /** @nullable */
  author?: string | null;
  format: string;
  extension: string;
  /** @nullable */
  fileSize?: string | null;
  createdAt: string;
}

export interface DownloadListResponse {
  downloads: Download[];
  total: number;
}

export interface StatsSummary {
  totalDownloads: number;
  totalVideos: number;
  totalAudios: number;
  totalImages: number;
  topPlatform: string;
  downloadsToday: number;
}

export interface PlatformStat {
  platform: string;
  count: number;
}

export type ListDownloadsParams = {
limit?: number;
offset?: number;
platform?: string;
mediaType?: string;
};

export type GetRecentDownloadsParams = {
limit?: number;
};