# lib/api-zod/src/generated/types/download.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediaDrop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Download {
  id: number;
  url: string;
  platform: string;
  mediaType: string;
  title: string;
  /** @nullable */
  thumbnail?: string | null;
  /** @nullable */
  author?: string | null;
  format: string;
  extension: string;
  /** @nullable */
  fileSize?: string | null;
  createdAt: Date;
}