# lib/api-zod/src/generated/types/downloadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediaDrop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadInput {
  url: string;
  platform: string;
  mediaType: string;
  title: string;
  /** @nullable */
  thumbnail?: string | null;
  /** @nullable */
  author?: string | null;
  /** Quality label selected */
  format: string;
  extension: string;
  /** @nullable */
  fileSize?: string | null;
}