# lib/api-zod/src/generated/types/downloadListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediaDrop API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Download } from './download';

export interface DownloadListResponse {
  downloads: Download[];
  total: number;
}