# lib/api-zod/src/generated/types/mediaFormat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediaDrop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MediaFormat {
  id: string;
  /** Human-readable quality label (e.g. "1080p HD", "MP3 320kbps") */
  label: string;
  /** video, audio, or image */
  type: string;
  extension: string;
  /** @nullable */
  quality?: string | null;
  /**
     * Estimated file size
     * @nullable
     */
  size?: string | null;
  /**
     * Direct download URL if available
     * @nullable
     */
  url?: string | null;
}