# lib/api-zod/src/generated/types/mediaInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediaDrop API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MediaFormat } from './mediaFormat';

export interface MediaInfo {
  url: string;
  /** Detected platform (youtube, instagram, tiktok, etc.) */
  platform: string;
  /** video, audio, image, or carousel */
  mediaType: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  /**
     * Duration in seconds
     * @nullable
     */
  duration?: number | null;
  /** @nullable */
  author?: string | null;
  formats: MediaFormat[];
}