# lib/api-zod/src/generated/types/parseUrlInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediaDrop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ParseUrlInput {
  /** The media URL to parse */
  url: string;
}