# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediaDrop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalDownloads: number;
  totalVideos: number;
  totalAudios: number;
  totalImages: number;
  topPlatform: string;
  downloadsToday: number;
}